GUI: if linked with Qt5 just call _exit() for exiting.
Under some desktops (at least KDE on OpenSUSE Tumbleweed plasma6 6.0.2), when built with qt5, either
exit() and qApp->exit() will loop and block, probably in some atexit() function. This only happens
when running on the local display, and not on a remote X11 display. Anyway, just call _exit() and be
xdone... This is a qt5 bug, because the same code built with qt6 can call qApp->exit() just fine.